+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## v1.0.1 - May 8, 2020
+
+Added script
+
+
+**Added:**
+- User Management
+- Secure root
+- Disable guest
+- change password expiration requirements
+- Enable firewall
+- Enable syn cookie protection
+- Disable IPv6
+- Disable IP Forwarding
+- Prevent IP Spoofing
+- Delete nmap and zenmap
+- Delete samba
+- Check for malware
+- Check for media files
+- Do updates
+
 ## v1.0.0 - May 8, 2020
 
 Initial release.